+++ dubbo-rpc/dubbo-rpc-http-invoker/src/main/java/org/apache/dubbo/rpc/protocol/httpinvoker/HttpInvokerProtocol.java
@@ -1,0 +1,225 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.dubbo.rpc.protocol.httpinvoker;
+
+import org.apache.dubbo.common.URL;
+import org.apache.dubbo.common.utils.StringUtils;
+import org.apache.dubbo.common.Version;
+import org.apache.dubbo.remoting.Constants;
+import org.apache.dubbo.remoting.http.HttpBinder;
+import org.apache.dubbo.remoting.http.HttpHandler;
+import org.apache.dubbo.remoting.http.HttpServer;
+import org.apache.dubbo.rpc.RpcContext;
+import org.apache.dubbo.rpc.RpcException;
+import org.apache.dubbo.rpc.protocol.AbstractProxyProtocol;
+import org.apache.dubbo.rpc.service.GenericService;
+import org.apache.dubbo.rpc.support.ProtocolUtils;
+
+import org.aopalliance.intercept.MethodInvocation;
+import org.springframework.remoting.RemoteAccessException;
+import org.springframework.remoting.httpinvoker.HttpComponentsHttpInvokerRequestExecutor;
+import org.springframework.remoting.httpinvoker.HttpInvokerProxyFactoryBean;
+import org.springframework.remoting.httpinvoker.HttpInvokerServiceExporter;
+import org.springframework.remoting.httpinvoker.SimpleHttpInvokerRequestExecutor;
+import org.springframework.remoting.support.RemoteInvocation;
+import org.springframework.remoting.support.RemoteInvocationFactory;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.net.HttpURLConnection;
+import java.net.SocketTimeoutException;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
+import static org.apache.dubbo.common.constants.CommonConstants.DEFAULT_TIMEOUT;
+import static org.apache.dubbo.common.constants.CommonConstants.RELEASE_KEY;
+import static org.apache.dubbo.common.constants.CommonConstants.TIMEOUT_KEY;
+import static org.apache.dubbo.remoting.Constants.DUBBO_VERSION_KEY;
+import static org.apache.dubbo.rpc.Constants.GENERIC_KEY;
+
+/**
+ * HttpInvokerProtocol
+ */
+public class HttpInvokerProtocol extends AbstractProxyProtocol {
+
+    public static final int DEFAULT_PORT = 80;
+
+    private final Map<String, HttpServer> serverMap = new ConcurrentHashMap<String, HttpServer>();
+
+    private final Map<String, HttpInvokerServiceExporter> skeletonMap = new ConcurrentHashMap<String, HttpInvokerServiceExporter>();
+
+    private HttpBinder httpBinder;
+
+    public HttpInvokerProtocol() {
+        super(RemoteAccessException.class);
+    }
+
+    public void setHttpBinder(HttpBinder httpBinder) {
+        this.httpBinder = httpBinder;
+    }
+
+    @Override
+    public int getDefaultPort() {
+        return DEFAULT_PORT;
+    }
+
+    @Override
+    protected <T> Runnable doExport(final T impl, Class<T> type, URL url) throws RpcException {
+        String addr = getAddr(url);
+        HttpServer server = serverMap.get(addr);
+        if (server == null) {
+            server = httpBinder.bind(url, new InternalHandler());
+            serverMap.put(addr, server);
+        }
+        final String path = url.getAbsolutePath();
+        skeletonMap.put(path, createExporter(impl, type));
+
+        final String genericPath = path + "/" + GENERIC_KEY;
+
+        skeletonMap.put(genericPath, createExporter(impl, GenericService.class));
+        return new Runnable() {
+            @Override
+            public void run() {
+                skeletonMap.remove(path);
+                skeletonMap.remove(genericPath);
+            }
+        };
+    }
+
+    private <T> HttpInvokerServiceExporter createExporter(T impl, Class<?> type) {
+        final HttpInvokerServiceExporter httpServiceExporter = new HttpInvokerServiceExporter();
+        httpServiceExporter.setServiceInterface(type);
+        httpServiceExporter.setService(impl);
+        try {
+            httpServiceExporter.afterPropertiesSet();
+        } catch (Exception e) {
+            throw new RpcException(e.getMessage(), e);
+        }
+        return httpServiceExporter;
+    }
+
+    @Override
+    @SuppressWarnings("unchecked")
+    protected <T> T doRefer(final Class<T> serviceType, final URL url) throws RpcException {
+        final String generic = url.getParameter(GENERIC_KEY);
+        final boolean isGeneric = ProtocolUtils.isGeneric(generic) || serviceType.equals(GenericService.class);
+
+        final HttpInvokerProxyFactoryBean httpProxyFactoryBean = new HttpInvokerProxyFactoryBean();
+        httpProxyFactoryBean.setRemoteInvocationFactory(new RemoteInvocationFactory() {
+            @Override
+            public RemoteInvocation createRemoteInvocation(MethodInvocation methodInvocation) {
+                RemoteInvocation invocation;
+                /*
+                  package was renamed to 'org.apache.dubbo' in v2.7.0, so only provider versions after v2.7.0 can
+                  recognize org.apache.xxx.HttpRemoteInvocation'.
+                 */
+                if (Version.isRelease270OrHigher(url.getParameter(RELEASE_KEY))) {
+                    invocation = new HttpRemoteInvocation(methodInvocation);
+                } else {
+                    /*
+                      The customized 'com.alibaba.dubbo.rpc.protocol.http.HttpRemoteInvocation' was firstly introduced
+                      in v2.6.3. The main purpose is to support transformation of attachments in HttpInvokerProtocol, see
+                      https://github.com/apache/dubbo/pull/1827. To guarantee interoperability with lower
+                      versions, we need to check if the provider is v2.6.3 or higher before sending customized
+                      HttpRemoteInvocation.
+                     */
+                    if (Version.isRelease263OrHigher(url.getParameter(DUBBO_VERSION_KEY))) {
+                        invocation = new com.alibaba.dubbo.rpc.protocol.http.HttpRemoteInvocation(methodInvocation);
+                    } else {
+                        invocation = new RemoteInvocation(methodInvocation);
+                    }
+                }
+                if (isGeneric) {
+                    invocation.addAttribute(GENERIC_KEY, generic);
+                }
+                return invocation;
+            }
+        });
+
+        String key = url.toIdentityString();
+        if (isGeneric) {
+            key = key + "/" + GENERIC_KEY;
+        }
+
+        httpProxyFactoryBean.setServiceUrl(key.replace("http-invoker","http"));
+        httpProxyFactoryBean.setServiceInterface(serviceType);
+        String client = url.getParameter(Constants.CLIENT_KEY);
+        if (StringUtils.isEmpty(client) || "simple".equals(client)) {
+            SimpleHttpInvokerRequestExecutor httpInvokerRequestExecutor = new SimpleHttpInvokerRequestExecutor() {
+                @Override
+                protected void prepareConnection(HttpURLConnection con,
+                                                 int contentLength) throws IOException {
+                    super.prepareConnection(con, contentLength);
+                    con.setReadTimeout(url.getParameter(TIMEOUT_KEY, DEFAULT_TIMEOUT));
+                    con.setConnectTimeout(url.getParameter(Constants.CONNECT_TIMEOUT_KEY, Constants.DEFAULT_CONNECT_TIMEOUT));
+                }
+            };
+            httpProxyFactoryBean.setHttpInvokerRequestExecutor(httpInvokerRequestExecutor);
+        } else if ("commons".equals(client)) {
+            HttpComponentsHttpInvokerRequestExecutor httpInvokerRequestExecutor = new HttpComponentsHttpInvokerRequestExecutor();
+            httpInvokerRequestExecutor.setReadTimeout(url.getParameter(TIMEOUT_KEY, DEFAULT_TIMEOUT));
+            httpInvokerRequestExecutor.setConnectTimeout(url.getParameter(Constants.CONNECT_TIMEOUT_KEY, Constants.DEFAULT_CONNECT_TIMEOUT));
+            httpProxyFactoryBean.setHttpInvokerRequestExecutor(httpInvokerRequestExecutor);
+        } else {
+            throw new IllegalStateException("Unsupported http protocol client " + client + ", only supported: simple, commons");
+        }
+        httpProxyFactoryBean.afterPropertiesSet();
+        return (T) httpProxyFactoryBean.getObject();
+    }
+
+    @Override
+    protected int getErrorCode(Throwable e) {
+        if (e instanceof RemoteAccessException) {
+            e = e.getCause();
+        }
+        if (e != null) {
+            Class<?> cls = e.getClass();
+            if (SocketTimeoutException.class.equals(cls)) {
+                return RpcException.TIMEOUT_EXCEPTION;
+            } else if (IOException.class.isAssignableFrom(cls)) {
+                return RpcException.NETWORK_EXCEPTION;
+            } else if (ClassNotFoundException.class.isAssignableFrom(cls)) {
+                return RpcException.SERIALIZATION_EXCEPTION;
+            }
+        }
+        return super.getErrorCode(e);
+    }
+
+    private class InternalHandler implements HttpHandler {
+
+        @Override
+        public void handle(HttpServletRequest request, HttpServletResponse response)
+                throws IOException, ServletException {
+            String uri = request.getRequestURI();
+            HttpInvokerServiceExporter skeleton = skeletonMap.get(uri);
+            if (!"POST".equalsIgnoreCase(request.getMethod())) {
+                response.setStatus(500);
+            } else {
+                RpcContext.getContext().setRemoteAddress(request.getRemoteAddr(), request.getRemotePort());
+                try {
+                    skeleton.handleRequest(request, response);
+                } catch (Throwable e) {
+                    throw new ServletException(e);
+                }
+            }
+        }
+
+    }
+
+}
