+++ dubbo-rpc/dubbo-rpc-http-invoker/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.Protocol
@@ -1,0 +1,1 @@
+http-invoker=org.apache.dubbo.rpc.protocol.httpinvoker.HttpInvokerProtocol
