--- dubbo-rpc/dubbo-rpc-jsonrpc/src/main/java/org/apache/dubbo/rpc/protocol/jsonrpc/JsonRpcProtocol.java
@@ -1,163 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.dubbo.rpc.protocol.jsonrpc;
-
-import org.apache.dubbo.common.URL;
-import org.apache.dubbo.remoting.http.HttpBinder;
-import org.apache.dubbo.remoting.http.HttpHandler;
-import org.apache.dubbo.remoting.http.HttpServer;
-import org.apache.dubbo.rpc.RpcContext;
-import org.apache.dubbo.rpc.RpcException;
-import org.apache.dubbo.rpc.protocol.AbstractProxyProtocol;
-
-import com.googlecode.jsonrpc4j.HttpException;
-import com.googlecode.jsonrpc4j.JsonRpcClientException;
-import com.googlecode.jsonrpc4j.JsonRpcServer;
-import com.googlecode.jsonrpc4j.spring.JsonProxyFactoryBean;
-import org.springframework.remoting.RemoteAccessException;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.net.SocketTimeoutException;
-import java.util.ArrayList;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-public class JsonRpcProtocol extends AbstractProxyProtocol {
-
-    public static final String ACCESS_CONTROL_ALLOW_ORIGIN_HEADER = "Access-Control-Allow-Origin";
-    public static final String ACCESS_CONTROL_ALLOW_METHODS_HEADER = "Access-Control-Allow-Methods";
-    public static final String ACCESS_CONTROL_ALLOW_HEADERS_HEADER = "Access-Control-Allow-Headers";
-
-    private final Map<String, HttpServer> serverMap = new ConcurrentHashMap<>();
-
-    private final Map<String, JsonRpcServer> skeletonMap = new ConcurrentHashMap<>();
-
-    private HttpBinder httpBinder;
-
-    public JsonRpcProtocol() {
-        super(HttpException.class, JsonRpcClientException.class);
-    }
-
-    public void setHttpBinder(HttpBinder httpBinder) {
-        this.httpBinder = httpBinder;
-    }
-
-    @Override
-    public int getDefaultPort() {
-        return 80;
-    }
-
-    private class InternalHandler implements HttpHandler {
-
-        private boolean cors;
-
-        public InternalHandler(boolean cors) {
-            this.cors = cors;
-        }
-
-        @Override
-        public void handle(HttpServletRequest request, HttpServletResponse response)
-                throws ServletException {
-            String uri = request.getRequestURI();
-            JsonRpcServer skeleton = skeletonMap.get(uri);
-            if (cors) {
-                response.setHeader(ACCESS_CONTROL_ALLOW_ORIGIN_HEADER, "*");
-                response.setHeader(ACCESS_CONTROL_ALLOW_METHODS_HEADER, "POST");
-                response.setHeader(ACCESS_CONTROL_ALLOW_HEADERS_HEADER, "*");
-            }
-            if ("OPTIONS".equalsIgnoreCase(request.getMethod())) {
-                response.setStatus(200);
-            } else if ("POST".equalsIgnoreCase(request.getMethod())) {
-
-                RpcContext.getContext().setRemoteAddress(request.getRemoteAddr(), request.getRemotePort());
-                try {
-                    skeleton.handle(request.getInputStream(), response.getOutputStream());
-                } catch (Throwable e) {
-                    throw new ServletException(e);
-                }
-            } else {
-                response.setStatus(500);
-            }
-        }
-
-    }
-
-    @Override
-    protected <T> Runnable doExport(T impl, Class<T> type, URL url) throws RpcException {
-        String addr = url.getIp() + ":" + url.getPort();
-        HttpServer server = serverMap.get(addr);
-        if (server == null) {
-            server = httpBinder.bind(url, new InternalHandler(url.getParameter("cors", false)));
-            serverMap.put(addr, server);
-        }
-        final String path = url.getAbsolutePath();
-        JsonRpcServer skeleton = new JsonRpcServer(impl, type);
-        skeletonMap.put(path, skeleton);
-        return () -> skeletonMap.remove(path);
-    }
-
-    @SuppressWarnings("unchecked")
-    @Override
-    protected <T> T doRefer(final Class<T> serviceType, URL url) throws RpcException {
-        JsonProxyFactoryBean jsonProxyFactoryBean = new JsonProxyFactoryBean();
-        jsonProxyFactoryBean.setServiceUrl(url.setProtocol("http").toIdentityString());
-        jsonProxyFactoryBean.setServiceInterface(serviceType);
-
-        jsonProxyFactoryBean.afterPropertiesSet();
-        return (T) jsonProxyFactoryBean.getObject();
-    }
-
-    @Override
-    protected int getErrorCode(Throwable e) {
-        if (e instanceof RemoteAccessException) {
-            e = e.getCause();
-        }
-        if (e != null) {
-            Class<?> cls = e.getClass();
-            if (SocketTimeoutException.class.equals(cls)) {
-                return RpcException.TIMEOUT_EXCEPTION;
-            } else if (IOException.class.isAssignableFrom(cls)) {
-                return RpcException.NETWORK_EXCEPTION;
-            } else if (ClassNotFoundException.class.isAssignableFrom(cls)) {
-                return RpcException.SERIALIZATION_EXCEPTION;
-            }
-        }
-        return super.getErrorCode(e);
-    }
-
-    @Override
-    public void destroy() {
-        super.destroy();
-        for (String key : new ArrayList<>(serverMap.keySet())) {
-            HttpServer server = serverMap.remove(key);
-            if (server != null) {
-                try {
-                    if (logger.isInfoEnabled()) {
-                        logger.info("Close jsonrpc server " + server.getUrl());
-                    }
-                    server.close();
-                } catch (Throwable t) {
-                    logger.warn(t.getMessage(), t);
-                }
-            }
-        }
-    }
-
-}
