--- dubbo-rpc/dubbo-rpc-jsonrpc/src/main/resources/META-INF/dubbo/internal/com.alibaba.dubbo.rpc.Protocol
@@ -1,1 +1,0 @@
-jsonrpc=org.apache.dubbo.rpc.protocol.jsonrpc.JsonRpcProtocol
